sync waypoint with word pin count buddy:qaiser

+++ src/DirectionResultSection/View/DirectionResultSectionController.h
@@ -33,7 +33,7 @@ namespace ExampleApp
                 std::vector<Search::SdkModel::SearchResultModel> m_lastAddedResults;
                 const Menu::View::IMenuReactionModel& m_menuReaction;
                 SearchResultPoi::View::ISearchResultPoiViewModel& m_searchResultPoiViewModel;
-                long m_wayPointCount;
+                int m_wayPointCount;
                 
                 Eegeo::Helpers::TCallback1<DirectionResultSectionController, const DirectionQueryResponseReceivedMessage&> m_directionResultReceivedHandler;
                 Eegeo::Helpers::TCallback1<DirectionResultSectionController, const Search::SearchQueryRemovedMessage&> m_directionQueryRemovedHandler;
